Made Significant
- Altered characters.js to treat var x/y as being centered on
character's feet. Removed click-to-die functionality for main character,
and attached him to the camera, which will now keep him centered as he
moves.

- Edited game.html to include new files level.js and levels\test.js.
Also resized screen from 800 by 600 to 1280 by 720.

- Added a level variable to gameengine.js, pointing to the current level
object it's drawing. Implemented x/y coordinates that alter where the
"origin" of the drawn canvas is in relation to the actual "origin", for
the purposes of simulating camera movement. Revamped draw() function to
cycle through and draw all tiles from the current level in addition to
all entities, and to draw them in the correct order. Entities calculated
to be outside the game window will no longer be drawn, to save
processing power.

- Added a new folder for tile spritesheets, as well as one actual
spritesheet for test purposes (basic.png).

- Renamed terrain.js to level.js, and significantly reworked it. Instead
of a simple background image, a Level object will now take in a
spritesheet of tiles and a 2D array indicating which to place and where
to place them on the screen. In other words, levels can now be
constructed tile by tile.

- Added a folder for storing levels, 2D arrays containing information on
where a Level object is to place its tiles. Also added a set of testing
arrays in test.js.

- Altered main.js to contain "static" variables for the screen
width/height, as well as to properly implement the new level system.

- Altered zombie.js (and character.js too, for that matter) to draw
themselves in relation to the x/y variables of the game engine, as part
of the moveable camera system.